Fix for multiple whitelisted patterns
The core loop originally would allow `X-Frame-Options` if the user specified any pattern that did not match the current URL. In other words, it would remove the header only when all patterns match.

This patch fixes that. At the same time, I moved some checks outside of the loop that don't depend on which regex is being checked.

I also removed the check for an empty array of regexes, on the grounds that the `for` loop will do the same check just as well.